Initial version of `FormatMoney` function.
We will want to have such a function for pretty-formating amount outputed by our backend (in format of cents). There is built-in really usefull `Number.prototype.toLocaleString` function that makes the hardest part for us.

The only possible problem is that it requires some kind of pollyfill on Safari Webkit, IE < 11 and most mobile browsers.